requirements.txt
websocket-client
requests
tuya-connector-python
dotenv